relational mapping

+++ python-object_relational_mapping/13-model_state_delete_a.py
@@ -1,32 +1,24 @@
 #!/usr/bin/python3
 """
-Module for deleting all the cities with 'a'.
+changes the name of a State object from the database hbtn_0e_6_usa
 """
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
 from sys import argv
-
+from sqlalchemy import (create_engine)
 from model_state import Base, State
+from sqlalchemy.orm import Session
 
-# Run only executed
+
 if __name__ == "__main__":
-
-    # Engine creation with mysql and mysqldb DBAPI
-    engine = create_engine("mysql+mysqldb://{}:{}@localhost:3306/{}"
-                           .format(argv[1], argv[2], argv[3]))
-
-    # Creating all classes in DB
+    engine = create_engine('mysql+mysqldb://{}:{}@localhost/{}'.format
+                           (argv[1], argv[2], argv[3]), pool_pre_ping=True)
     Base.metadata.create_all(engine)
 
-    # Creating Session and its instance
-    Session = sessionmaker(bind=engine)
-    session = Session()
+    session = Session(engine)
 
-    # Deleting everything with a
-    for state in session.query(State).filter(State.name.like('%a%')).all():
-        session.delete(state)
+    for state in session.query(State).order_by(State.id).all():
+        for c in state.name:
+            if c == 'a':
+                session.delete(state)
+                break
     session.commit()
-
-    # Closing the session
-    if session:
-        session.close()
+    session.close()
